DRY move content related resources
