+++ lib/lims-laboratory-app/laboratory/sample/swap_samples_resource.rb
@@ -1,24 +1,15 @@
 require 'lims-api/core_action_resource'
 require 'lims-api/struct_stream'
 require 'lims-laboratory-app/laboratory/sample/swap_samples'
+require 'lims-laboratory-app/laboratory/move_content_resource_shared'
 
 module Lims::LaboratoryApp
   module Laboratory
     class Sample
       class SwapSamplesResource < Lims::Api::CoreActionResource
 
-        def filtered_attributes
-          super.tap do |attributes|
-            attributes[:parameters] = attributes[:parameters].map.each do |element|
-              element.mash do |k,v|
-                case k
-                when "resource" then ["resource_uuid", @context.uuid_for(v)]
-                else [k,v]
-                end
-              end
-            end
-          end
-        end
+        include MoveContentResourceShared
+
       end
     end
   end
